Matlab code for iteration and results
I iterated N2 and operating B to calculate the cost and the efficiency of the transformer. May have mistakes in the coding. 60 different transformer designs are present in the results. I only printed the results which were also present in the picture. I will choose a transformer considering core mass cost and efficiency.